feat(worker): add single worker service with redundancy
